Move mapped classes with database table to another file tables.py
