+++ bot/database/__init__.py
@@ -17,4 +17,4 @@ DATABASE_URL = config("DATABASE_URL")
 ENGINE = create_engine(DATABASE_URL, echo=True)
 session_factory = sessionmaker(bind=ENGINE)
 SESSION = scoped_session(session_factory)
-BASE = declarative_base()
+Base = declarative_base()
